refactor: 예약 서비스 통합
