--- src/main/java/roomescape/service/AdminReservationService.java
@@ -1,66 +1,0 @@
-package roomescape.service;
-
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import roomescape.domain.Reservation;
-import roomescape.domain.ReservationTime;
-import roomescape.domain.Theme;
-import roomescape.exception.NotFoundException;
-import roomescape.repository.ReservationRepository;
-import roomescape.repository.ReservationTimeRepository;
-import roomescape.repository.ThemeRepository;
-
-import java.time.LocalDate;
-import java.util.List;
-
-@Service
-@Transactional(readOnly = true)
-public class AdminReservationService {
-
-    private final ReservationRepository reservationRepository;
-    private final ReservationTimeRepository reservationTimeRepository;
-    private final ThemeRepository themeRepository;
-    private final ReservationValidator reservationValidator;
-
-    public AdminReservationService(ReservationRepository reservationRepository, ReservationTimeRepository reservationTimeRepository, ThemeRepository themeRepository, ReservationValidator reservationValidator) {
-        this.reservationRepository = reservationRepository;
-        this.reservationTimeRepository = reservationTimeRepository;
-        this.themeRepository = themeRepository;
-        this.reservationValidator = reservationValidator;
-    }
-
-    public List<Reservation> findAll() {
-        return reservationRepository.findAll();
-    }
-
-    @Transactional
-    public Reservation create(String name, LocalDate date, Long timeId, Long themeId) {
-        ReservationTime time = findReservationTime(timeId);
-        reservationValidator.validateAlreadyReserved(date, timeId, themeId);
-        Theme theme = findTheme(themeId);
-
-        Reservation reservation = new Reservation(null, name, date, time, theme);
-        return save(reservation);
-    }
-
-    @Transactional
-    public void delete(Long id) {
-        reservationRepository.delete(id);
-    }
-
-    private Reservation save(Reservation reservation) {
-        Long id = reservationRepository.insert(reservation);
-        return reservationRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("생성된 예약을 찾을 수 없습니다."));
-    }
-
-    private ReservationTime findReservationTime(Long timeId) {
-        return reservationTimeRepository.findBy(timeId)
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 예약 시간입니다."));
-    }
-
-    private Theme findTheme(Long themeId) {
-        return themeRepository.findBy(themeId)
-                .orElseThrow(() -> new NotFoundException("존재하지 않는 테마입니다."));
-    }
-}
